chore!: rebrand as onionpipe
Trademark disputes can leave a project like this one vulnerable to
seizure and takeover by well-funded parties.

Even though "onion" is a vegetable and "grok" is jargon-speak, best to
steer clear. Those who have the most gold will write the rules.

So this project is rebranding as onionpipe. Which is probably a better
name and brand for what this is anyway.

BREAKING CHANGE: rename of all packages & artifacts.